test: fold test_sigv4 into test_s3 and test_oauth into test_google
app/sigv4.py has exactly one consumer chain: auth.resolve_sigv4 -> routers/s3.py. app/oauth.py
backs /oauth2/token, /_mock/credentials and google.py's service-account email — the credentials
are Google-shaped and only Google's router reads them. Neither warranted its own file.

app/pagination.py was checked too and STAYS: it is used by six routers (atlassian, github, google,
notion, slack, and both GraphQL resolvers), which makes it the most cross-vendor module in the
repo, not a Linear/Fireflies helper.

Caught while merging: hoisting the botocore imports had put them ABOVE
`pytest.importorskip("botocore")`, which would turn a skip into a hard ImportError on a machine
without botocore. Guard order asserted in this commit's verification.

787 test functions before and after, none lost. pytest 860 passed; extras-gated 28 passed.
